Un-refactor ECSV tests to test_serialize [skip ci]
